Add stiles for the tooltip and verify wallets styles
